levelset: fix evaluation of segment information for points that lie on the segmentation
Due to numerical errors, points that lie on the surface may have a
distance slightly grater than zero. Therefore, a small tolerance is
now used when checking if a point lies on the segmentation.

Moreover, if a point lies on the segmentation, we assume that it is
on the outer side of the surface, this means that the distance is a
positive zero and that the gradient is set equal to the surface normal.